Add error handling

+++ entrypoint.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
-set -eo pipefail
+# Purpose: Detecting Hardware Errors
+# Author: Vivek Gite <vivek@nixcraft.com>
+# Last updated on : 28-Aug-2007
+# -----------------------------------------------
+
+set -o pipefail
 
 if [[ -z "${SRC_PATH}" ]]; then
   echo "SRC_PATH environment variable is missing. Cannot proceed."
@@ -45,21 +50,33 @@ git config --global user.email "${GITHUB_ACTOR}@users.noreply.github.com"
 
 echo "Copying \"${SRC_REPO_NAME}/${SRC_PATH}\" and pushing it to ${GITHUB_REPOSITORY}"
 
-
 git clone --branch ${SRC_BRANCH} --single-branch --depth 1 https://${GITHUB_TOKEN}@github.com/${SRC_REPO}.git
-rm -rf ${SRC_REPO_NAME}/.git
+if [ "$?" -ne 0 ]; then
+    echo >&2 "Cloning '$SRC_REPO' failed"
+    exit 1
+fi
+rm -rf ${SRC_REPO_NAME}/.git # TODO remove every file that matches a filter. See issue #1.
 
 git clone --branch ${DST_BRANCH} --single-branch --depth 1 https://${GH_PAT}@github.com/${DST_REPO}.git
+if [ "$?" -ne 0 ]; then
+    echo >&2 "Cloning '$DST_REPO' failed"
+    exit 1
+fi
 
+mkdir -p ${DST_REPO_NAME}/${DIR} || exit "$?"
+cp -rf ${SRC_REPO_NAME}/${SRC_PATH} ${DST_REPO_NAME}/${DST_PATH} || exit "$?"
 
-mkdir -p ${DST_REPO_NAME}/${DIR}
-cp -rf ${SRC_REPO_NAME}/${SRC_PATH} ${DST_REPO_NAME}/${DST_PATH}
+cd ${DST_REPO_NAME} || exit "$?"
 
-cd ${DST_REPO_NAME}
+if [ -z "$(git status --porcelain)" ]; then
+    # Working directory is clean
+    echo "No changes detected "
+else
+    # Uncommitted changes
+    git add -A
+    git commit --message "Update \"${SRC_PATH}\" from \"${GITHUB_REPOSITORY}\""
 
-git add -A
-git commit --message "Update \"${SRC_PATH}\" from \"${GITHUB_REPOSITORY}\""
-
-git push -u origin ${DST_BRANCH}
+    git push -u origin ${DST_BRANCH}
+fi
 
 echo "Copying complete 👌"
